[WK2][EFL]: Invalidate the ProcessLauncher when the process is terminated before it has finished launching https://bugs.webkit.org/show_bug.cgi?id=117865
Patch by Sergio Correia <sergio.correia@openbossa.org> on 2013-06-21
Reviewed by Christophe Dumez.

Original patch from Carlos Garcia Campos for the Gtk port.

* UIProcess/Launcher/efl/ProcessLauncherEfl.cpp:
(WebKit::ProcessLauncher::terminateProcess): If process is still
launching, just invalidate the launcher. Reset the process
identifier after killing the process.

Canonical link: https://commits.webkit.org/136004@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@151828 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/Launcher/efl/ProcessLauncherEfl.cpp
@@ -114,9 +114,15 @@ void ProcessLauncher::launchProcess()
 
 void ProcessLauncher::terminateProcess()
 {
+    if (m_isLaunching) {
+        invalidate();
+        return;
+    }
+
     if (!m_processIdentifier)
         return;
     kill(m_processIdentifier, SIGKILL);
+    m_processIdentifier = 0;
 }
 
 void ProcessLauncher::platformInvalidate()
